Move some modules to programmable-tokens-test and -offchain
