+++ src/examples/aiken/haskell/aiken-example.cabal
@@ -81,10 +81,10 @@ test-suite aiken-example-test
   other-modules:  Wst.Aiken.Test
   build-depends:
     , aiken-example
-    , base                          >=4.14.0
+    , base                         >=4.14.0
     , cardano-api
     , containers
     , programmable-tokens-onchain
-    , programmable-tokens-test-lib
+    , programmable-tokens-test
     , tasty
     , tasty-hunit
