+++ src/examples/aiken/haskell/test/Wst/Aiken/Test.hs
@@ -8,8 +8,8 @@ import Cardano.Api qualified as C
 import Data.Functor (void)
 import Data.Map qualified as Map
 import Paths_aiken_example qualified as Pkg
-import SmartTokens.Core.Scripts (ScriptTarget (Debug, Production))
-import SmartTokens.Test qualified as Test
+import ProgrammableTokens.Test qualified as Test
+import SmartTokens.Core.Scripts (ScriptTarget (Production))
 import Test.Tasty (TestTree, testGroup)
 import Test.Tasty.HUnit (Assertion, assertEqual, testCase)
 import Wst.Aiken.Blueprint (Blueprint (..))
@@ -20,7 +20,7 @@ tests = testGroup "unit tests"
   [ testCase "load blueprint" loadBlueprint
   , testCase "deserialise script" deserialiseScript
   , testGroup "emulator" [
-      testCase "register" (Test.mockchainSucceedsWithTarget _ action)
+      testCase "register" (Test.mockchainSucceedsWithTarget Production registerAikenPolicy)
   ]
   ]
 
@@ -43,3 +43,6 @@ loadExample = do
   Pkg.getDataFileName "data/aiken-scripts.json"
     >>= Blueprint.loadFromFile
     >>= either fail pure
+
+registerAikenPolicy :: Monad m => m ()
+registerAikenPolicy = pure ()
