+++ src/examples/regulated-stablecoin/lib/Wst/Offchain/Query.hs
@@ -18,7 +18,6 @@ module Wst.Offchain.Query(
 
 import Cardano.Api qualified as C
 import Control.Lens qualified as L
-import Control.Monad ((>=>))
 import Control.Monad.Error.Lens (throwing_)
 import Control.Monad.Except (MonadError)
 import Control.Monad.Reader (MonadReader, asks)
@@ -27,59 +26,25 @@ import Convex.Class (MonadBlockchain (queryNetworkId), MonadUtxoQuery,
                      utxosByPaymentCredential)
 import Convex.PlutusLedger.V1 (transCredential, transPolicyId,
                                unTransStakeCredential)
-import Convex.Scripts (fromHashableScriptData)
-import Convex.Utxos (UtxoSet, toApiUtxo)
-import Data.Aeson (FromJSON, ToJSON)
-import Data.Aeson qualified as JSON
 import Data.List (sortOn)
-import Data.Map qualified as Map
-import Data.Maybe (listToMaybe, mapMaybe)
-import Data.OpenApi.Schema (ToSchema (..))
-import Data.OpenApi.Schema qualified as Schema
-import Data.OpenApi.SchemaOptions qualified as SchemaOptions
+import Data.Maybe (listToMaybe)
 import Data.Ord (Down (..))
-import Data.Typeable (Typeable)
 import GHC.Exts (IsList (..))
-import GHC.Generics (Generic)
-import PlutusTx qualified
+import ProgrammableTokens.OffChain.Orphans ()
+import ProgrammableTokens.OffChain.UTxODat (UTxODat (..), extractUTxO,
+                                            extractUtxoNoDatum, fromOutput)
 import SmartTokens.Contracts.IssuanceCborHex (IssuanceCborHex (..))
 import SmartTokens.Types.ProtocolParams (ProgrammableLogicGlobalParams)
 import SmartTokens.Types.PTokenDirectory (BlacklistNode, DirectorySetNode (..))
 import Wst.AppError (AsProgrammableTokensError (..))
-import Wst.JSON.Utils qualified as JSON
 import Wst.Offchain.Env (DirectoryEnv (..), HasDirectoryEnv (directoryEnv),
                          HasTransferLogicEnv (transferLogicEnv),
                          TransferLogicEnv (tleBlacklistSpendingScript),
                          blacklistNodePolicyId, directoryNodePolicyId,
                          issuanceCborHexPolicyId,
                          programmableTokenMintingScript, protocolParamsPolicyId)
-import Wst.Orphans ()
 
 -- TODO: We should probably filter the UTxOs to check that they have the correct NFTs
-
-{-| Unspent transaction output with 'TxIn', 'TxOut' and an inline datum
--}
-data UTxODat era a =
-  UTxODat
-    { uIn    :: C.TxIn
-    , uOut   :: C.TxOut C.CtxUTxO era
-    , uDatum :: a
-    }
-  deriving stock (Eq, Show, Generic)
-
--- | Aeson options for the UTxODat type. Used to derive JSON instances and ToSchema
-utxoDatOptions :: JSON.Options
-utxoDatOptions = JSON.customJsonOptions 1
-
-instance (C.IsCardanoEra era, ToJSON a) => ToJSON (UTxODat era a) where
-  toJSON = JSON.genericToJSON utxoDatOptions
-  toEncoding = JSON.genericToEncoding utxoDatOptions
-
-instance (C.IsCardanoEra era, FromJSON a, C.IsShelleyBasedEra era) => FromJSON (UTxODat era a) where
-  parseJSON = JSON.genericParseJSON utxoDatOptions
-
-instance (Typeable a, ToSchema a, Typeable era) => ToSchema (UTxODat era a) where
-  declareNamedSchema = Schema.genericDeclareNamedSchema (SchemaOptions.fromAesonOptions utxoDatOptions)
 
 {-| Find all UTxOs that make up the registry
 -}
@@ -152,20 +117,6 @@ programmableLogicOutputs :: forall era env m. (MonadReader env m, HasDirectoryEn
 programmableLogicOutputs = do
   asks (C.PaymentCredentialByScript . C.hashScript . C.PlutusScript C.PlutusScriptV3 . dsProgrammableLogicBaseScript . directoryEnv)
     >>= fmap (extractUtxoNoDatum @era) . utxosByPaymentCredential
-
-fromOutputNoDatum :: forall era. (C.IsBabbageBasedEra era) => C.TxIn -> C.TxOut C.CtxUTxO era -> Maybe (UTxODat era ())
-fromOutputNoDatum uIn uOut@(L.preview (L._TxOut . L._3 . L._TxOutDatumInline) >=> fromHashableScriptData -> Just ()) = Just UTxODat{uIn, uOut, uDatum = ()}
-fromOutputNoDatum uIn uOut = Just $ UTxODat{uIn, uOut, uDatum = ()}
-
-extractUtxoNoDatum :: forall era b. (C.IsBabbageBasedEra era) => UtxoSet C.CtxUTxO b -> [UTxODat era ()]
-extractUtxoNoDatum = mapMaybe (uncurry fromOutputNoDatum) . Map.toList . C.unUTxO . toApiUtxo @era
-
-fromOutput :: forall era a. (PlutusTx.FromData a, C.IsBabbageBasedEra era) => C.TxIn -> C.TxOut C.CtxUTxO era -> Maybe (UTxODat era a)
-fromOutput uIn uOut@(L.preview (L._TxOut . L._3 . L._TxOutDatumInline) >=> fromHashableScriptData -> Just uDatum) = Just UTxODat{uIn, uOut, uDatum}
-fromOutput _ _ = Nothing
-
-extractUTxO :: forall era a b. (PlutusTx.FromData a, C.IsBabbageBasedEra era) => UtxoSet C.CtxUTxO b -> [UTxODat era a]
-extractUTxO = mapMaybe (uncurry fromOutput) . Map.toList . C.unUTxO . toApiUtxo @era
 
 extractValue :: C.IsBabbageBasedEra era => C.TxOut C.CtxUTxO era -> C.Value
 extractValue = L.view $ L._TxOut . L._2 . L._TxOutValue
